Fix Listing issues

+++ project_crates/lib/backend/storageAccess.dart
@@ -1,5 +1,5 @@
 import 'dart:typed_data';
-
+import 'package:uuid/uuid.dart';
 import 'package:path/path.dart' as Path;
 import 'package:firebase_storage/firebase_storage.dart';
 import 'dart:io';
@@ -8,10 +8,12 @@ import 'package:path_provider/path_provider.dart';
 
 class StorageAccess {
   FirebaseStorage storage = FirebaseStorage.instance;
+  var uuid = Uuid();
 
   Future<String> uploadFile(File file) async {
+    String randomString = uuid.v4();
     StorageReference storageReference =
-        storage.ref().child('Listing/${Path.basename(file.path)}');
+        storage.ref().child('Listing/$randomString${Path.basename(file.path)}');
     StorageUploadTask uploadTask = storageReference.putFile(file);
     await uploadTask.onComplete;
     print('File uploaded');
@@ -36,6 +38,7 @@ class StorageAccess {
   }
 
   Future<void> deleteListingImage(String url) async {
+    if (url == null) return;
     StorageReference storageRef = await storage.getReferenceFromUrl(url);
     await storageRef.delete();
     print("Storage Image deleted");
